generator/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
httpx  # ⭐️ Retriever API와 통신하기 위해 필수
langchain_openai
langchain_qdrant
langchain_core